rename service plugin to static
