+++ api/v1/upstream.proto
@@ -21,7 +21,7 @@ message Upstream {
     // Name of the upstream. Names must be unique and follow the following syntax rules:
     // One or more lowercase rfc1035/rfc1123 labels separated by '.' with a maximum length of 253 characters.
     string name = 1;
-    // Type indicates the type of the upstream. Examples include service<!--(TODO)-->, kubernetes<!--(TODO)-->, and [aws](../plugins/aws.md)
+    // Type indicates the type of the upstream. Examples include static<!--(TODO)-->, kubernetes<!--(TODO)-->, and [aws](../plugins/aws.md)
     // Types are defined by the plugin<!--(TODO)--> associated with them.
     string type = 2;
     // Connection Timeout tells gloo to set a timeout for unresponsive connections created to this upstream.
