+++ docs/getting_started/docker/1.md
@@ -21,7 +21,7 @@ an upstream with `glooctl`
 ```
 cat << EOF | glooctl upstream create -f -
 name: petstore
-type: service
+type: static
 spec:
   hosts:
   # gateway ip for the docker network
@@ -35,7 +35,7 @@ You should see a new upstream has been created.
 +----------+---------+--------+----------+
 |   NAME   |  TYPE   | STATUS | FUNCTION |
 +----------+---------+--------+----------+
-| petstore | service |        |          |
+| petstore | static |        |          |
 +----------+---------+--------+----------+
 ```
 
@@ -47,7 +47,7 @@ glooctl upstream get
 +----------+---------+----------+-------------+
 |   NAME   |  TYPE   |  STATUS  |  FUNCTION   |
 +----------+---------+----------+-------------+
-| petstore | service | Accepted | addPet      |
+| petstore | static | Accepted | addPet      |
 |          |         |          | deletePet   |
 |          |         |          | findPetById |
 |          |         |          | findPets    |
@@ -104,7 +104,7 @@ spec:
     port: 9000
 status:
   state: Accepted
-type: service
+type: static
 ```
 
 #### Step 4
