+++ docs/plugins/static.md
@@ -3,15 +3,15 @@
 
 #### Description
 
-The [Service Plugin for Gloo](https://github.com/solo-io/gloo-plugins/tree/master/service) is a basic plugin enabling
+The [Static Plugin for Gloo](https://github.com/solo-io/gloo/tree/master/pkg/coreplugins/static) is a basic plugin enabling
 routing to an upstream which is simply a list of host/port combinations for a single service. 
-A typical use case for defining `service` upstreams is to route to external services, or route to a service whose upstream
+A typical use case for defining `static` upstreams is to route to external services, or route to a service whose upstream
 type is not yet supported by an existing Gloo plugin.
 
 
 #### Upstream Spec Configuration
 
-The **Upstream Type** for service upstreams is `service`. 
+The **Upstream Type** for service upstreams is `static`. 
 
 The [upstream spec](../v1/upstream.md#v1.Upstream) for Service Upstreams has the following structure:
 
@@ -25,12 +25,12 @@ hosts:
 
 | Field | Type |  Description |
 | ----- | ---- |  ----------- |
-| hosts | []Host |  a list of Hosts to which routes for this service should connect. **at least one required** |
+| hosts | []Host |  a list of Hosts to which routes for this upstream should connect. **at least one required** |
 
 A Host has the following structure:
 
-| addr | string |  an IP or Hostname for the service. **required** |
-| port | int | the port on which to reach the service
+| addr | string |  an IP or Hostname for the upstream. **required** |
+| port | int | the port on which to reach the upstream
 
 #### Example Service Upstream
 
@@ -44,5 +44,5 @@ spec:
     port: 8080
   - addr: some-host.example.com
     port: 1234
-type: service
+type: static
 ```
