+++ example/nats/nats-upstream.yaml
@@ -1,4 +1,4 @@
-type: service
+type: static
 name: nats-streaming
 metadata:
   annotations:
