+++ pkg/coreplugins/static/plugin.go
@@ -1,4 +1,4 @@
-package service
+package static
 
 import (
 	"net"
