+++ pkg/coreplugins/static/spec.go
@@ -1,4 +1,4 @@
-package service
+package static
 
 import (
 	"github.com/pkg/errors"
