+++ pkg/plugins/fake/spec.go
@@ -3,15 +3,15 @@ package fake
 import (
 	"github.com/gogo/protobuf/types"
 	"github.com/solo-io/gloo/pkg/api/types/v1"
-	"github.com/solo-io/gloo/pkg/coreplugins/service"
+	"github.com/solo-io/gloo/pkg/coreplugins/static"
 )
 
-type UpstreamSpec = service.UpstreamSpec
+type UpstreamSpec = static.UpstreamSpec
 
 func DecodeUpstreamSpec(generic v1.UpstreamSpec) (UpstreamSpec, error) {
-	return service.DecodeUpstreamSpec(generic)
+	return static.DecodeUpstreamSpec(generic)
 }
 
 func EncodeUpstreamSpec(spec UpstreamSpec) *types.Struct {
-	return service.EncodeUpstreamSpec(spec)
+	return static.EncodeUpstreamSpec(spec)
 }
