+++ test/kube_e2e/core_service_plugin_test.go
@@ -6,7 +6,7 @@ import (
 	. "github.com/onsi/ginkgo"
 	"github.com/pborman/uuid"
 	"github.com/solo-io/gloo/pkg/api/types/v1"
-	"github.com/solo-io/gloo/pkg/coreplugins/service"
+	"github.com/solo-io/gloo/pkg/coreplugins/static"
 	. "github.com/solo-io/gloo/test/helpers"
 )
 
@@ -19,9 +19,9 @@ var _ = Describe("Core Service Plugin", func() {
 		BeforeEach(func() {
 			_, err := gloo.V1().Upstreams().Create(&v1.Upstream{
 				Name: helloService,
-				Type: service.UpstreamTypeService,
-				Spec: service.EncodeUpstreamSpec(service.UpstreamSpec{
-					Hosts: []service.Host{
+				Type: static.UpstreamTypeService,
+				Spec: static.EncodeUpstreamSpec(static.UpstreamSpec{
+					Hosts: []static.Host{
 						{
 							Addr: helloService,
 							Port: servicePort,
