+++ test/kube_e2e/events_test.go
@@ -9,7 +9,7 @@ import (
 
 	. "github.com/onsi/ginkgo"
 	"github.com/solo-io/gloo/pkg/api/types/v1"
-	"github.com/solo-io/gloo/pkg/coreplugins/service"
+	"github.com/solo-io/gloo/pkg/coreplugins/static"
 	"github.com/solo-io/gloo/pkg/log"
 	. "github.com/solo-io/gloo/test/helpers"
 )
@@ -22,9 +22,9 @@ var _ = Describe("Event matcher route type", func() {
 		BeforeEach(func() {
 			_, err := gloo.V1().Upstreams().Create(&v1.Upstream{
 				Name: upstreamForEvents,
-				Type: service.UpstreamTypeService,
-				Spec: service.EncodeUpstreamSpec(service.UpstreamSpec{
-					Hosts: []service.Host{
+				Type: static.UpstreamTypeService,
+				Spec: static.EncodeUpstreamSpec(static.UpstreamSpec{
+					Hosts: []static.Host{
 						{
 							Addr: upstreamForEvents,
 							Port: servicePort,
