+++ test/kube_e2e/ssl_virtualhost_test.go
@@ -7,7 +7,7 @@ import (
 
 	. "github.com/onsi/ginkgo"
 	"github.com/solo-io/gloo/pkg/api/types/v1"
-	"github.com/solo-io/gloo/pkg/coreplugins/service"
+	"github.com/solo-io/gloo/pkg/coreplugins/static"
 	. "github.com/solo-io/gloo/test/helpers"
 	kubev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
@@ -49,9 +49,9 @@ var _ = Describe("SNI VirtualService", func() {
 			Must(err)
 			_, err = gloo.V1().Upstreams().Create(&v1.Upstream{
 				Name: helloService,
-				Type: service.UpstreamTypeService,
-				Spec: service.EncodeUpstreamSpec(service.UpstreamSpec{
-					Hosts: []service.Host{
+				Type: static.UpstreamTypeService,
+				Spec: static.EncodeUpstreamSpec(static.UpstreamSpec{
+					Hosts: []static.Host{
 						{
 							Addr: helloService,
 							Port: servicePort,
